% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogwrangler.R
\name{ogcode}
\alias{ogcode}
\title{go from english/welsh language name to instance code}
\usage{
ogcode(name, entity = NULL, language = "english")
}
\description{
go from english/welsh language name to instance code
}
