% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogwrangler.R
\name{ogfind}
\alias{ogfind}
\title{pass in e.g. leicester to search for what that might be
TODO improve}
\usage{
ogfind(name, nresults = 10)
}
\description{
pass in e.g. leicester to search for what that might be
TODO improve
}
