% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogwrangler.R
\name{ogmetric}
\alias{ogmetric}
\title{Get measure data for a set of geographical entities}
\usage{
ogmetric(key, from, metric = "pop2018")
}
\arguments{
\item{key}{code for measures; pass "*" for all}

\item{from}{rules for what to interpret codes in key as}

\item{metric}{metric to return}
}
\description{
Get measure data for a set of geographical entities
}
\examples{

# Get population for English admin regions
ogmetric("*", "e.reg", "pop2018")
}
