% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogwrangler.R
\name{ogname}
\alias{ogname}
\title{translates code into english or welsh; if suffix true append some kinda descriptor}
\usage{
ogname(code, language = "english", suffix = F)
}
\description{
translates code into english or welsh; if suffix true append some kinda descriptor
}
