% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogwrangler.R
\name{ogwrangle}
\alias{ogwrangle}
\title{Transform count or proportion data from one set of geography codes to another}
\usage{
ogwrangle(x, key, from, to, by = "pop2018", measure = "count")
}
\arguments{
\item{x}{numerical data to wrangle}

\item{key}{code associated with each element of x}

\item{from}{rules for what to interpret codes in key as}

\item{to}{rules for what to translate codes in key to}

\item{by}{unit over which to distribute parts of x}

\item{measure}{"count" if x is count data, "proportion" if x is proportion data}
}
\value{
Transformed data
}
\description{
This function takes data that are mapped to one set of geography codes and redistributes
those data to another set of geography codes, proportionally to either the population
or the surface area of the regions referred to by the geography codes.
}
\details{
\code{from} and \code{to} can either be a single string specifying the entity being mapped
from/to, or a named character vector where the name specifies a regex for matching a subset
of the geography codes in \code{key} and the value specifies the entity to map the codes in
\code{from}.
}
\examples{

# Map prevalence data from English admin
# regions to NHS England regions
data = fread(
"RGN19CD	prev
E12000001	4.6
E12000002	4.9
E12000003	3.2
E12000004	4.2
E12000005	5.1
E12000006	6.2
E12000007	11.9
E12000008	4.8
E12000009	2.3")
data[, ogwrangle(prev, RGN19CD,
                 "e.reg", "e.nhser20",
                 "pop2018", "proportion")];

# Map data from countries to regional health boards.
# This only provides a "best guess" to count data
# within these smaller regions, since of course
# the actual distribution of these count data
# within countries is not known!
data = fread(
"country    n
E92000001   1000
N92000002   100
S92000003   250
W92000004   200")
data[, ogwrangle(
          n, country, "country",
          c("E" = "e.nhser20", "N" = "n.hsct",
            "S" = "s.hb", "W" = "w.lhb"),
          "pop2018", "count")];

}
